Move PersonDTO to dto package
