+++ condo-manager-adapter/condo-manager-adapter-web/src/main/kotlin/com/ikm/condomanager/adapter/web/dto/PersonDTO.kt
@@ -1,4 +1,4 @@
-package com.ikm.condomanager.adapter.web.controller
+package com.ikm.condomanager.adapter.web.dto
 
 import com.ikm.condomanager.domain.PersonId
