+++ condo-manager-adapter/condo-manager-adapter-web/src/test/kotlin/com/ikm/condomanager/adapter/web/controller/PersonControllerTest.kt
@@ -3,6 +3,7 @@ package com.ikm.condomanager.adapter.web.controller
 import com.ikm.condomanager.adapter.web.converter.convertToPerson
 import com.ikm.condomanager.adapter.web.converter.convertToPersonDTO
 import com.ikm.condomanager.adapter.web.converter.mergeToPerson
+import com.ikm.condomanager.adapter.web.dto.PersonDTO
 import com.ikm.condomanager.domain.Person
 import com.ikm.condomanager.domain.PersonId
 import com.ikm.condomanager.usecase.person.CreatePersonUseCase
